+++ lino_welfare/modlib/courses/models.py
@@ -506,10 +506,11 @@ class CourseRequest(dd.Model):
             return CourseOffer.objects.filter(content=content)
         return CourseOffer.objects.all()
     
-    def before_state_change(self,ar,kw,old,new):
+    def before_state_change(self,ar,old,new):
         if new.name in ('passed','award','failed','aborted'):
             if not self.date_ended:
                 self.date_ended = datetime.date.today()
+        super(CourseRequest,self).before_state_change(ar,old,new)
       
     def get_row_permission(self,ar,state,ba):
         if not super(CourseRequest,self).get_row_permission(ar,state,ba):